fix: 修改路由请求hook调用方式
