+++ packages/server/package.json
@@ -20,7 +20,8 @@
     "dayjs": "^1.11.13",
     "express": "^4.19.2",
     "jsonwebtoken": "^9.0.2",
-    "multer": "1.4.5-lts.1"
+    "multer": "1.4.5-lts.1",
+    "rc-dropdown": "4.0.1"
   },
   "devDependencies": {
     "@types/bcryptjs": "^2.4.6",
