--- packages/server/src/store/66c4a9b4008a6f92455fab0b/1.txt
@@ -1,1 +1,0 @@
-12345
